Updated WDManager
Mild stuff. Fixed the issue with creating a window at various resolutions not working
was just an issue with comparison operators being wrong.
Working on fixing resizing the window after it is displayed and it is not working.
I think that the problem may be that the window is throwing events for resizing
and currently there is no event manager.